Fix fields on schema being shared via reference.
There is also a few other quality of life improvements. For example
collection_type on CollectionField might be a type or instance (instance
allows us to define name and description per individual object). Also
data on fields has been made consistent (notably SchemaField.data
behaves like a dictionary and not as a list).