Add new schema output/input DocumentField
It's mongo-ish type, any key and any value will do. Different to python
it accepts anything that is json-serializible (ie. key as dictionary will
do). The limitation is json representation, key or value is invalid form
of json, you have to override it with your own field definition and
normalisation.